requirements.txt
agate==1.6.3
altair==4.2.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.8
attrs==22.1.0
babel==2.11.0
backcall==0.2.0
backports-zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blis==0.7.9
cachetools==4.2.2
catalogue==2.0.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
cinemagoer==2022.2.11
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.7
dbt-core==1.3.1
dbt-extractor==0.4.1
dbt-snowflake==1.3.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
fiona==1.8.22
fonttools==4.38.0
future==0.18.2
geojson==2.5.0
geopandas==0.9.0
gitdb==4.0.9
gitpython==3.1.29
gmaps==0.9.0
greenlet==2.0.1
hologram==0.0.15
htbuilder==0.6.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.2
imdbpy==2022.7.9
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
iteration-utilities==0.11.0
jaraco-classes==3.2.3
jedi==0.18.1
jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.4.4
jupyter-console==6.4.4
jupyter-core
openai
langchain
faiss-cpu
sentence_transformers
catboost
